logica de dinero implementada. varias actualizaciones y seeder de consolas

+++ database/migrations/2024_05_01_171413_create_consolas_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nombre');
             $table->string('descripcion');
             $table->string('ruta_imagen');
-            $table->integer('money');
+            $table->float('money');
             $table->timestamps();
         });
     }
